+++ librouteros/api.py
@@ -26,6 +26,16 @@ class Api:
         self.protocol.writeSentence(cmd, *words)
         yield from self._readResponse()
 
+    def rawCmd(self, cmd, *words):
+        """
+        Call Api with given command and raw words.
+        End user is responsible to properly format each api word argument.
+        :param cmd: Command word. eg. /ip/address/print
+        :param args: Iterable with optional plain api arguments.
+        """
+        self.protocol.writeSentence(cmd, *words)
+        yield from self._readResponse()
+
     def _readSentence(self):
         """
         Read one sentence and parse words.
@@ -69,3 +79,38 @@ class Api:
         >>> '/ip/address/print'
         """
         return pjoin('/', *path).rstrip('/')
+
+    def path(self, *path):
+        return Path(
+                path='',
+                api=self,
+                ).join(*path)
+
+
+class Path:
+    """Represents absolute command path."""
+
+    def __init__(self, path, api):
+        self.path = path
+        self.api = api
+
+    def select(self, key, *other):
+        keys = (key,) + other
+        return Query(path=self, keys=keys, api=self.api)
+
+    def __str__(self):
+        return self.path
+
+    def __repr__(self):
+        return "<{module}.{cls} {path!r}>".format(
+            module=self.__class__.__module__,
+            cls=self.__class__.__name__,
+            path=self.path,
+        )
+
+    def join(self, *path):
+        """Join current path with one or more path strings."""
+        return Path(
+                api=self.api,
+                path=pjoin('/', self.path, *path).rstrip('/'),
+                )